split core models between core and chamber_of_deputies models
